requirements.txt
pandas
yfinance
plotly
jupyter
feast
requests
beautifulsoup4
azure-storage-blob
azure-storage-file-datalake
azure-identity
python-dotenv